+++ website/src/routes/(docs)/methods/api/(methods)/getDirtyInput/properties.ts
@@ -1,0 +1,105 @@
+import type { PropertyProps } from '~/components';
+
+export const properties: Record<string, PropertyProps> = {
+  TSchema: {
+    modifier: 'extends',
+    type: {
+      type: 'custom',
+      name: 'Schema',
+      href: '/core/api/Schema/',
+    },
+  },
+  TFieldPath: {
+    modifier: 'extends',
+    type: {
+      type: 'union',
+      options: [
+        {
+          type: 'custom',
+          name: 'RequiredPath',
+          href: '/core/api/RequiredPath/',
+        },
+        'undefined',
+      ],
+    },
+    default: 'undefined',
+  },
+  form: {
+    type: {
+      type: 'custom',
+      name: 'BaseFormStore',
+      href: '/core/api/BaseFormStore/',
+      generics: [
+        {
+          type: 'custom',
+          name: 'TSchema',
+        },
+      ],
+    },
+  },
+  config: {
+    type: {
+      type: 'union',
+      options: [
+        {
+          type: 'custom',
+          name: 'GetFormDirtyInputConfig',
+          href: '../GetFormDirtyInputConfig/',
+        },
+        {
+          type: 'custom',
+          name: 'GetFieldDirtyInputConfig',
+          href: '../GetFieldDirtyInputConfig/',
+          generics: [
+            {
+              type: 'custom',
+              name: 'TSchema',
+            },
+            {
+              type: 'custom',
+              name: 'RequiredPath',
+            },
+          ],
+        },
+        'undefined',
+      ],
+    },
+  },
+  result: {
+    type: {
+      type: 'union',
+      options: [
+        {
+          type: 'custom',
+          name: 'DeepPartial',
+          href: '/core/api/DeepPartial/',
+          generics: [
+            {
+              type: 'custom',
+              name: 'PathValue',
+              href: '/core/api/PathValue/',
+              generics: [
+                {
+                  type: 'custom',
+                  name: 'v.InferInput',
+                  href: 'https://valibot.dev/api/InferInput/',
+                  generics: [
+                    {
+                      type: 'custom',
+                      name: 'TSchema',
+                    },
+                  ],
+                },
+                {
+                  type: 'custom',
+                  name: 'TFieldPath',
+                },
+              ],
+            },
+          ],
+        },
+        'undefined',
+      ],
+    },
+  },
+};