+++ website/src/routes/(docs)/methods/api/(types)/GetFieldDirtyInputConfig/properties.ts
@@ -1,0 +1,44 @@
+import type { PropertyProps } from '~/components';
+
+export const properties: Record<string, PropertyProps> = {
+  TSchema: {
+    modifier: 'extends',
+    type: {
+      type: 'custom',
+      name: 'Schema',
+      href: '/core/api/Schema/',
+    },
+  },
+  TFieldPath: {
+    modifier: 'extends',
+    type: {
+      type: 'custom',
+      name: 'RequiredPath',
+      href: '/core/api/RequiredPath/',
+    },
+  },
+  path: {
+    type: {
+      type: 'custom',
+      name: 'ValidPath',
+      href: '/core/api/ValidPath/',
+      generics: [
+        {
+          type: 'custom',
+          name: 'v.InferInput',
+          href: 'https://valibot.dev/api/InferInput/',
+          generics: [
+            {
+              type: 'custom',
+              name: 'TSchema',
+            },
+          ],
+        },
+        {
+          type: 'custom',
+          name: 'TFieldPath',
+        },
+      ],
+    },
+  },
+};